forecaster/requirements.txt
accelerate==1.13.0
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
chronos-forecasting==2.2.2
click==8.3.1
cmdstanpy==1.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
cython==3.2.4
einops==0.8.2
fastapi==0.135.1
filelock==3.25.0
fonttools==4.61.1
fsspec==2026.2.0
h11==0.16.0
holidays==0.92
huggingface-hub==0.36.2
idna==3.11
importlib-resources==6.5.2
jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pandas==3.0.1
patsy==1.0.2
pillow==12.1.1
pmdarima==2.1.1
polars==1.39.3
polars-runtime-32==1.39.3
prophet==1.3.0
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyyaml==6.0.3
regex==2026.2.28
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
six==1.17.0
stanio==0.5.1
starlette==0.52.1
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==4.57.6
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
